Split listener into two classes and refactor common code
Splits Listener into two classes: HandleMessage and StompClient
Utils has been moved into its own seperate file to keep the LOC count
down.

The tests were left alone (except for fixing up imports) as the split
was completed to ensure nothing regressed.
Next _get_last_run was created to unify some common code and fix a bug,
making mocking much easier. Normally this would be done in its own
commit, but I didn't realise the split work hadn't been comitted. So I'm
going to be cheeky and slip it in here.

If you get here from a bisect, I apolgoise: _get_last_run was updated to
take an experiment and run number (rather than just a run number) as
data_ready could potentially bump us straight to version 1 of a run
across two different instruments if they same run number turned up
across them

This
